% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_year.R
\name{assert_cy}
\alias{assert_cy}
\title{Determine duplicated country-years}
\usage{
assert_cy(df)
}
\arguments{
\item{df}{DataFrame}
}
\value{
If there are duplicated country-years in the given data
frame, \code{assert_cy} will throw an error listing the offending
countries. Otherwise, \code{NULL} is returned.
}
\description{
Assert that there are no duplicated country-years in the given
DataFrame.
}
