% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_peace_yrs}
\alias{calc_peace_yrs}
\title{Calculate peace years}
\usage{
calc_peace_yrs(years, incidence)
}
\arguments{
\item{years}{NumericVector of years}

\item{incidence}{NumericVector or LogicalVector of same size as
\code{year} indicating conflict incidence}
}
\description{
Given a vector of years and a matching length vector of conflict
incidence, calculate the number of peace years since the last civil
conflict or the coding start, identified as the first element.
}
\details{
This calculates the number of years of peace since either
the last incidence of civil conflict or the coding start given
by the first element in the \code{years} vector.\preformatted{Note, peace years are calculated based on V-Dem countries,
which means that gwid changes are not recorded as changes in
the fundamental status of a country. This only affects Serbia
in 2006 and Czech Republic in 1993.

Furthermore, the coding starts are also taken from V-Dem, which
includes non-sovergeign polities such as colonial
territories. In the `merge.R` we attempt to therefore backtrack
conflicts on the basis of territory location (ex: Eritrea),
even though the final analysis will only be done on independent
countries according to GW.
}
}
\examples{
years <- 1900:1910
ongoing <- sample(0:1, length(years), replace = TRUE)

calc_peace_yrs(years, ongoing)

}
