% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_changes}
\alias{collapse_changes}
\title{Collapse identical observations}
\usage{
collapse_changes(x)
}
\arguments{
\item{x}{Vector of any type}
}
\description{
Given a vector of any type, return the grouping indices for
consecutively unchanged elements.
}
\examples{
collapse_changes(c(rep(1, 3), 4, rep(3, 2)))

}
