% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{consecutive}
\alias{consecutive}
\title{Find consecutive elements}
\usage{
consecutive(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Given a NumericVector, return the grouping indices for consecutive
elements.
}
\examples{
consecutive(c(1900:1905, 1908:1910))

}
