% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_year.R
\name{dbg_info}
\alias{dbg_info}
\title{Merging stats}
\usage{
dbg_info(df)
}
\arguments{
\item{df}{DataFrame}
}
\description{
Extremely simply debugging function that does nothing more than
simply print out the number of conflict episodes, countries, and
country-years in a given data frame. Completely useless for any
generic task.
}
