% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explode}
\alias{explode}
\title{Explode a data frame!}
\usage{
explode(df, from = NULL, to = NULL)
}
\arguments{
\item{df}{DataFrame}

\item{from}{NumericVector, starting value of sequence}

\item{to}{NumericVector, end value of sequence}
}
\value{
A \code{data.frame} with the added column, \code{sequence}, with the
values defined by \code{from:to}.
}
\description{
Given a \code{data.frame}, repeat each row according to the inclusive
length of the sequence defined by \code{from} and \code{to}.
}
\details{
This is only used to expand a collapsed \code{data.frame} to a
full country-year time series. In other words, given a
\code{data.frame} where each row represents a range of years,
\code{explode} will duplicate each row for each year in the range,
if given as the \code{from} and \code{to} arguments.
}
\examples{
df <- data.frame(x = 1:3)
explode(df, from = rep(1, 3), to = 2:4)

}
