% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{extract}
\alias{extract}
\title{Extract parameters}
\usage{
extract(x, pars = NULL)
}
\arguments{
\item{x}{DataFrame with the class \code{posterior}.}

\item{pars}{CharacterVector of target parameters. If the parameter
is a vector or matrix, the function will extract all matching
values.}
}
\description{
Extracts posteriors from a \code{posterior} object created using
\code{read_post}.
}
