% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt}
\alias{fmt}
\title{Pretty print numerics}
\usage{
fmt(x, digits = 2)
}
\arguments{
\item{x}{NumericVector}

\item{digits}{Number of decimal digits to round to.}
}
\value{
CharacterVector
}
\description{
This function formats a numeric value by rounding to the specified
number of digits before then calling \code{prettyNum}.
}
\examples{
fmt(10000.1243)
fmt(9999.999)

}
