% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.ordinal}
\alias{is.ordinal}
\title{Ordinal Values}
\usage{
is.ordinal(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Checks a numeric vector if all the values can be considered
ordinal. This is equivalent to saying that the function checks if a
given numeric vector can be coerced to an integer vector without
loss.
}
\details{
This is only intended for NumericVectors. For all other
object types, the function will return \code{FALSE}.
}
\examples{
x <- 1:3
is.ordinal(x)

y <- c(1:3, 0.5)
is.ordinal(y)

}
