% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{is.posterior}
\alias{is.posterior}
\title{Posterior class}
\usage{
is.posterior(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Boolean
}
\description{
Checks whether an object inherits from the \code{posterior} class as
returned by \link{read_post}.
}
