% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Simple wrapper around the default arguments to \code{\link[=scale]{scale()}} which
transforms a numeric matrix according to \code{(x - mean(x)) / sd(x)}.
}
\examples{
normalize(1:10)

}
