% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_pars}
\alias{overlay_pars}
\title{Overlay parameter plots}
\usage{
overlay_pars(pars = NULL, legend_labels = NULL, ...)
}
\arguments{
\item{pars}{Parameters to extract}

\item{legend_labels}{Optional CharacterVector labeling each
posterior object in the plot's legend.}

\item{...}{Variable number of posterior objects}
}
\description{
Given a list of posterior objects, \code{overlay_pars} extracts the
specified parameters and plots the parameter estimates (point
estimate + 80 percent CIs) as a single overlayed plot.
}
