% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partial}
\alias{partial}
\title{Partial application}
\usage{
partial(fn, ...)
}
\arguments{
\item{fn}{Function}

\item{...}{Arguments to be applied}
}
\description{
Produce a new function by binding the provided arguments to \code{fn}.
}
\examples{
f <- partial(sum, na.rm = TRUE)
f(c(1, NA, 2, 3))

}
