% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dens}
\alias{plot_dens}
\title{Posterior density plot}
\usage{
plot_dens(m, n = 50, groups = NULL)
}
\arguments{
\item{m}{Posterior matrix}

\item{n}{Number of posterior draws to be used}

\item{groups}{Grouping factor to separate observations}
}
\description{
Plots the density curves for the first \code{n} posterior draws from the
posterior matrix \code{m}.
}
