% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_summarise}
\alias{post_summarise}
\title{Summarise posterior distribution}
\usage{
post_summarise(x, ...)
}
\arguments{
\item{x}{NumericMatrix or \code{stanfit} object}

\item{...}{Arguments passed to methods}
}
\description{
Calculates quantile intervals from a posterior distribution
represented either by a NumericMatrix where rows are posterior
draws or extracted directly from a \code{posterior} object.
}
\details{
By default, \code{post_summarise} will calculate 2.5\% and 97.5\%
quantile intervals in addition to the posterior median.
}
