% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{read_post}
\alias{read_post}
\title{Load posterior object}
\usage{
read_post(file)
}
\arguments{
\item{file}{Target file. Can be a compressed csv (\code{csv.gz}).}
}
\value{
DataFrame with the additional class, \code{posterior}.
}
\description{
Reads a posterior file and creates a \code{posterior} object.
}
