% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_stats}
\alias{summary_stats}
\title{Calculate Summary Statistics}
\usage{
summary_stats(x, vars = colnames(x))
}
\arguments{
\item{x}{DataFrame}

\item{vars}{CharacterVector. Calculate statistics for only a subset
of columns.}
}
\value{
A data frame with the following descriptive statistics: N,
mean, standard deviation, min, and max.
}
\description{
Calculates summary statistics for all numeric columns in a data
frame.
}
\examples{
df <- data.frame(x = letters[1:3], y = 1:3, z = rnorm(3))
summary_stats(df)

}
