% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tfplot}
\alias{tfplot}
\title{Tie Fighter Plot}
\usage{
tfplot(df)
}
\arguments{
\item{df}{DataFrame containing \code{country_name}, \code{year}, and the
summary statistics for a selected parameter (\code{codelow},
\code{median}, \code{codehigh})}
}
\description{
Plots posterior intervals for each country-year for a summarised
parameter.
}
\details{
Unlike the other plotting functions, \code{tfplot} does not
accept as input the raw posteriors. This is because it's a
relatively inflexible function that requires \code{country_name} and
\code{year} to already be matched for each observation.
}
