% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.utils.R
\docType{package}
\name{thesis.utils}
\alias{thesis.utils}
\alias{thesis.utils-package}
\title{thesis.utils}
\description{
Utility functions for executive constraints and civil conflict
analysis. The majority of exported functions are fairly specific
and are not meant to be particularly useful outside the context of
this analysis.
}
