% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_idx}
\alias{to_idx}
\title{Create index}
\usage{
to_idx(x, ...)
}
\arguments{
\item{x}{Vector of any type}

\item{...}{Additional arguments passed to \code{\link[=factor]{factor()}}}
}
\value{
NumericVector indexing the original values of \code{x}.
}
\description{
Create a NumericVector of index positions from the unique
categories of a given vector, \code{x}.
}
\details{
This is useful for one thing only: creating the input data
list for Stan.\preformatted{For partially pooled intercepts in Stan we need the index
positions of each unique category associated with a given
observation.

For example, matching a random intercept for N years involves
estimating 1:N varying intercepts that need to be properly
indexed in the final regression. See `stan/model.stan` and
`R/model.R`.
}
}
\examples{
years <- rep(1900:1905, times = 2)
to_idx(years)

}
