% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pars}
\alias{plot_pars}
\title{Parameter plot}
\usage{
plot_pars(x, ...)
}
\arguments{
\item{m}{Posterior matrix where each column is a separate parameter}

\item{labels}{Optional CharacterVector of x-axis labels.}
}
\description{
Plots posterior median plus 68/95 quantile intervals for the given
parameters.
}
